% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_thmodels.R
\name{combine_thmodels}
\alias{combine_thmodels}
\title{Concatenates thmodels.  Warning: quadratic runtime for long lists.}
\usage{
combine_thmodels(name = "", thmodels = list())
}
\arguments{
\item{name}{# user-friendly name, used for titles of plots}

\item{thmodels}{# a list of thmodels to be combined}
}
\value{
thmodel
}
\description{
Concatenates thmodels.  Warning: quadratic runtime for long lists.
}
\examples{
tm <- combine_thmodels("eNV50kWh", list(eNV200ac50kWh, eNV200noac50kWh))
}
