% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params.R
\name{default_params}
\alias{default_params}
\title{default_params: sets thmodel params to default values}
\usage{
default_params(m, ot = NULL)
}
\arguments{
\item{m}{a thmodel}

\item{ot}{an ocv_tbl, estimated from voltage behaviour (est_ocv())}
}
\value{
a thmodel with all required params set
}
\description{
I'm relying on Kubal (2022, https://doi.org/10.1016/j.jpowsour.2022.231864)
for an estimation of the effective activation energy in the temperature
dependence (as estimated by Arrhenius' equation) of the effective resistance
of the NCM811 (or perhaps NCM632) cells in the Nissan packs under test.
An activation energy in the range from -27 to -31 kJ/mol corresponds to a
slope in the range [-3250, -3730] of the log-resistance of the pack (for
purposes of a Joule-heating estimation) as a function of temperature.  I
have validated -3500 as a plausible fit to the temperature-dependence of
the (inaccurately-predicted) effective resistance of the pack.  I don't
have anywhere near enough data to form a more accurate estimate of the
arrhenius_resistance parameter in my model of the pack's thermal behaviour.
}
\details{
I'm relying on fragments of a manufacturer's spec sheet posted online by a
retailer of the CALB L300N137B 3.7 volt 3.7V 137ah lithium nmc battery cell,
for a charge-discharge curve at 0.2C which I interpolated to derive a first
estimate of the pack's OCV as a function of its SOC. My intention is to
refine this estimate using observed data -- to gain a more accurate estimate
of the effective pack resistance as a function of SOC, and also perhaps to
add a second rate constant (as in the Butler-Volmer equation) to model
voltage drops (and thus Joule heating) more accurately than is possible with
a single-resistor model.  But: 150 mOhm (as derived from the charge-discharge
curve at 70% SOC and 0.2C) is currently our default estimate of the effective
pack resistance at 25 °C -- at all SOC.

Polarisation shifts near the cells' electrodes are highly entropic, so will
cause non-negligible (but reversible) heating and cooling whenever the pack
voltage shifts.  I roughly model this thermal gain/loss as a
polarisation_energy * delta_v.  By observation (of plots generated by
plot_fit()) these entropic shifts cause the 50kWh pack to cool by about two
degrees at the beginning of every fastcharge session, and to heat by a couple
of degrees at the end of that session.  Because this is a reversible heat,
its effects while driving are not dramatic over short time intervals -- but
they do add up to a significant cooling for each 10V drop in the pack's OCV.
Entropic heating has been well-studied in the technical literature, and it
arises from several sources -- some of which are highly SOC-dependent.  I
thus don't expect my single-parameter modelling of this phenomenon to be very
accurate. My initial empirical investigation suggests that a few kJ/V
is a reasonable first estimate for the 50 kWh pack.  The cells in the 24 kWh
may have a different electrolyte with different salt content, but as a first
approximation I'm expecting it to have half the polarisation_energy of the 50
kWh pack.

As a first estimate of heat_capacity: The 96 cells in my aftermarket 50kWh
pack weigh 2.13 kg apiece.  If the solvent in their electrolyte is DME
(1,2-dimethoxyethane), it has a heat capacity of 192 J/mol.K, i.e. about 2.1
J/gK.  If the solvent is ethylene carbonate, then it would have a lower heat
capacity: 134 J/mol.K = 1.6 J/gK. The primary salt in the electrolyte might
be lithium hexafluorophosphate (152 J/mol.K, 151 g/mol, 1.0 J/gK).  Its
presence increases the density of the electrolyte, but its lower heat content
would decrease the per-gram heat capacity of a 1:1 mix of DME and EC to 1.4
J/gK, if it is in a 1M solution. The casings of the prismatic cells may be
polyethylene (at 2.0 J/gK). The module casings seem to be steel (at 0.5
J/gK).  The shell of the pack is aluminium (at 0.9 J/gK). As a round number,
the total weight of the pack is 220 kg.  Its heat capacity might thus be 300
to 400 kJ/K.  An alternative estimate from the literature: a 60 Ah NMC811
pouch cell was estimated as having a specific heat capacity of 1.1 J/gK
(see https://www.sciencedirect.com/science/article/pii/S1452398125002512),
suggesting that the heat capacity of the VIVNE 50kWh pack might be
250 kJ/K.  Our initial estimate of heat_capacity is thus 300 kJ/K for the
50kWh pack.  The 24kWh pack has approximately the same weight, so we use the
same initial estimate of 300 kJ/K for its heat_capacity.  Our temperature
and voltage models are imprecise, and the pack resistance parameters are
almost entirely collinear with the pack's heat capacity (with the exception
of the reversible heat of polarisation); but our best-fits suggest a
heat capacity of approximately 250 kJ/K.

1 GID = 80 Wh, as confirmed experimentally in https://cthombor.
wpcomstaging.com/50kwh-upgrade-to-my-e-nv200/50kwh-upgrade-to-my-
24kwh-2014-nissan-e-nv200-part-3-estimation-of-usable-kwh/.
The dashboard-reported SOC is adjusted to have a 0-point when the
pack has about 10% of its nominal capacity.  The LeafSpy-reported
SOC has no allowance for the "reserve"; and the vehicle will
show a "turtle" on the dashboard when the LeafSpy-reported SOC
is approximately 10% of its maximum value.
A cursory examination of plot_gids(eNV50kWh) reveals that the
firmware for the BMS in this 50 kWh pack produces an estimate
for SOC that is often quite inaccurate when the pack's SOC is between
1/3 and 1/2 of its nominal capacity.  The BMS in Alastair's 50 kWh
pack might, or might not, have a similar issue given that there
is such little diversity in the operating conditions of
eNVa.50kWh.2025.  The OEM BMS in my 24kWh e-NV200 shows an excellent
fit to SOC = 9.9 + 0.309 * GID / SOH, over the quite varied operating
conditions of eNV24kWh. The BMS in my 24kWh Leaf shows an excellent
fit to SOC = 11.2 + 0.299 GID / SOH over the operating
conditions of Leaf24kWh_2019 (which had no SOC < 45).  These fits motivate
my default value of 10 for gids_reserve.
}
\examples{
default_params(new_thmodel())
}
