% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV200ac24kWh_2023}
\alias{eNV200ac24kWh_2023}
\title{eNV200ac24kWh_2023: a thmodel object of logs from 2023}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV200ac24kWh_2023
}
\description{
eNV200ac24kWh_2023: a thmodel object of logs from 2023
}
\keyword{datasets}
