% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV200ac24kWh_2024}
\alias{eNV200ac24kWh_2024}
\title{eNV200ac24kWh_2024: a thmodel object of logs from 2024}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV200ac24kWh_2024
}
\description{
eNV200ac24kWh_2024: a thmodel object of logs from 2024
}
\keyword{datasets}
