% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV200ac24kWh_2025}
\alias{eNV200ac24kWh_2025}
\title{eNV200ac24kWh_2025: a thmodel object, with provenance.  The fields are
as defined in thmodel.R.  The logdata is a tibble with more than 3600 rows
and more than 150 columns.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV200ac24kWh_2025
}
\description{
eNV200ac24kWh_2025: a thmodel object, with provenance.  The fields are
as defined in thmodel.R.  The logdata is a tibble with more than 3600 rows
and more than 150 columns.
}
\keyword{datasets}
