% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV200ac24kWh_2025mdy}
\alias{eNV200ac24kWh_2025mdy}
\title{eNV200ac24kWh_2025mdy: a thmodel object of logs from 2025 when
my LeafSpy was configured to store timestamps in USonian format.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV200ac24kWh_2025mdy
}
\description{
eNV200ac24kWh_2025mdy: a thmodel object of logs from 2025 when
my LeafSpy was configured to store timestamps in USonian format.
}
\keyword{datasets}
