% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV24kWh}
\alias{eNV24kWh}
\title{eNV24kWh: a very large thmodel object containing logs from my eNV200
prior to the replacement of its 24kWh pack.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV24kWh
}
\description{
eNV24kWh: a very large thmodel object containing logs from my eNV200
prior to the replacement of its 24kWh pack.
}
\keyword{datasets}
