% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV50kWh}
\alias{eNV50kWh}
\title{Datasets for testing ThermalEV}
\format{
A thmodel object.  Fields in the logdata include
  * {date_time}{timestamp in POSIXct format}
  * {elv}{elevation in meters above sea level, an integer}
  * {soc}{state of charge, an integer in the range 0:1e6}
  * {pack_volts}{pack voltage, at 0.1 volt precision}
  * {pack_amps}{pack amperage, at 0.001 amp precision}
  * {pack_t1_c}{temperature at surface of module 1, at 0.1 K precision}
  * {pack_t2_c}{temperature at surface of module 2, at 0.1 K precision}
  * {pack_t3_c}{uniformly NA in our version of LeafSpy}
  * {pack_t4_c}{temperature at surface of module 4, at 0.1 K precision}
  * {odo_km}{odometer reading, if car in Drive mode, otherwise NA}
}
\source{
{thmodel} R package.
}
\usage{
eNV50kWh
}
\description{
eNV50kWh: a very large thmodel object, with provenance.  The fields are
as defined in thmodel.R.  The logdata is a tibble with 27922 rows
and 185 columns.  It is constructed by concatenating (using rbind)
the logdata in eNV200noac50kWh and eNV200ac50kWh.
}
\keyword{datasets}
