% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNVa24kWh_2024}
\alias{eNVa24kWh_2024}
\title{eNVa24kWh_2024: a thmodel object of Alastair's logs from his 24kWh
eNV200, in 2024.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNVa24kWh_2024
}
\description{
eNVa24kWh_2024: a thmodel object of Alastair's logs from his 24kWh
eNV200, in 2024.
}
\keyword{datasets}
