% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNVa50kWh_2025}
\alias{eNVa50kWh_2025}
\title{eNVa50kWh_2025: a thmodel object of Alastair's logs from his 50kWh
eNV200.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNVa50kWh_2025
}
\description{
eNVa50kWh_2025: a thmodel object of Alastair's logs from his 50kWh
eNV200.
}
\keyword{datasets}
