% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model7}
\alias{fit_model7}
\title{fit_model7: uses optim() to find a best-fit of seven non-resistive params,
with a MSE criterion}
\usage{
fit_model7(
  m = NULL,
  arrhenius_resistance = NA,
  heat_capacity = NA,
  entropic_heat = NA,
  polarisation_irr = NA,
  lambda_module_to_ambient = NA,
  lambda_module_AC_to_ambient = NA,
  fan_power = NA,
  COP = NA,
  effective_pack_resistance = NA,
  packr85 = NA,
  tafel_slope = NA,
  lambda_polarisation = NA,
  arrhenius_tafel = NA,
  iter_count = 4,
  min_segment_length = 20,
  fixed_parameters = c(F, F, F, F, F, F, F),
  trace = 1,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{arrhenius_resistance}{in K, not optimised}

\item{heat_capacity}{in kJ/K}

\item{entropic_heat}{in kJ/V}

\item{polarisation_irr}{dimensionless, irreversible heat of polarisation}

\item{lambda_module_to_ambient}{in hours}

\item{lambda_module_AC_to_ambient}{in hours}

\item{fan_power}{in W}

\item{COP}{coefficient of heatpump performance, dimensionless}

\item{effective_pack_resistance}{in mOhms, not optimised}

\item{packr85}{in mOhms, not optimised}

\item{tafel_slope}{large-C polarisation response, not optimised}

\item{lambda_polarisation}{in s, decay of tafel polarisation, not optimised}

\item{arrhenius_tafel}{in K, Arrhenius coefficient for tafel, not optimised}

\item{iter_count}{controls convergence on predicted temps}

\item{min_segment_length}{shorter sequences of samples are ignored}

\item{fixed_parameters}{length-6 Boolean vector, reduces dimension of opt}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}

\item{from_date}{starting date/time (for a time-restricted optimisation)}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{x}{parameter list}
}
\value{
modified thmodel, with best-fit predictions and parameters
}
\description{
fit_model7: uses optim() to find a best-fit of seven non-resistive params,
with a MSE criterion
}
\examples{
m <- fit_model7(thmodels = list("eNV200ac50kWh"))
fm: local fcn, interface to predict_temp(), for use by optim()

Side effect: updates thmodel `m` in the calling environment

}
