% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ocv.R
\name{fit_ocv}
\alias{fit_ocv}
\title{fit_ocv: uses optim() to find a best-fit, with MSE criterion}
\usage{
fit_ocv(thmodels = list())
}
\arguments{
\item{thmodels}{a list of names of thmodels}

\item{x}{parameter list}
}
\value{
modified ocv_model, with best-fit ocv_tbl
}
\description{
fit_ocv: uses optim() to find a best-fit, with MSE criterion
}
\examples{
m <- fit_ocv(list("eNV200ac50kWh"))
fm: local fcn, for use by optim()

Side effect: updates ocv_model om in the calling environment

}
