% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_r_to_ocv.R
\name{fit_r_to_ocv}
\alias{fit_r_to_ocv}
\title{fit_r_to_ocv: uses optim() to find a best-fit pack resistance to an ocv_model}
\usage{
fit_r_to_ocv(om = NULL, iter_count = 4, fit_packr85 = F, trace = 1)
}
\arguments{
\item{om}{an ocv_model, or a thmodel (for convenience)}

\item{iter_count}{controls convergence on predicted temps}

\item{fit_packr85}{optimise packr85 separately from effective_pack_resistance}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}

\item{x}{parameter list}
}
\value{
modified ocv_model, with best-fit resistive params
}
\description{
fit_r_to_ocv: uses optim() to find a best-fit pack resistance to an ocv_model
}
\examples{
om <- fit_r_to_ocv(eNV50kWh)
fom: local fcn, for use by optim()

Side effect: updates ocv_model om in the calling environment

}
