% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leaf24kWh}
\alias{Leaf24kWh}
\title{Leaf24kWh: all my LeafSpy records from a 2013 24kWh Leaf}
\format{
A thmodel object
}
\usage{
Leaf24kWh
}
\description{
Leaf24kWh: all my LeafSpy records from a 2013 24kWh Leaf
}
\keyword{datasets}
