% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leaf24kWh_2019}
\alias{Leaf24kWh_2019}
\title{Leaf24kWh_2019: a thmodel object of logs from my 24kWh Leaf in 2019}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
Leaf24kWh_2019
}
\description{
Leaf24kWh_2019: a thmodel object of logs from my 24kWh Leaf in 2019
}
\keyword{datasets}
