% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leaf24kWh_2021}
\alias{Leaf24kWh_2021}
\title{Leaf24kWh_2021: a thmodel object of logs from my 24kWh Leaf in 2021}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
Leaf24kWh_2021
}
\description{
Leaf24kWh_2021: a thmodel object of logs from my 24kWh Leaf in 2021
}
\keyword{datasets}
