% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Leaf24kWh_2022}
\alias{Leaf24kWh_2022}
\title{Leaf24kWh_2022: a thmodel object of logs from my 24kWh Leaf in 2022}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
Leaf24kWh_2022
}
\description{
Leaf24kWh_2022: a thmodel object of logs from my 24kWh Leaf in 2022
}
\keyword{datasets}
