% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocv_tbl_50}
\alias{ocv_tbl_50}
\title{ocv_tbl_50: a lookup tibble of SOC onto OCV, empirically derived for the
50kWh pack (with extremal values from the cell manufacturer's specs)}
\format{
51 obs of 2 variables
}
\source{
{thmodel} R package.
}
\usage{
ocv_tbl_50
}
\description{
ocv_tbl_50: a lookup tibble of SOC onto OCV, empirically derived for the
50kWh pack (with extremal values from the cell manufacturer's specs)
}
\keyword{datasets}
