% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_12V.R
\name{plot_12V}
\alias{plot_12V}
\title{plot_12V: plots distance, speed, LVDC volts and amps; geolocates in title}
\usage{
plot_12V(m, from_date = NULL, to_date = NULL, from_idx = NULL, to_idx = NULL)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
plot_12V: plots distance, speed, LVDC volts and amps; geolocates in title
}
\examples{
plot_12V(predict_temp())
plot_12V(predict_temp(), 1, 10)
}
