% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot pack temp, predicted pack temp, ambient temp, kW (smoothed)}
\usage{
plot_fit(m, from_date = NULL, to_date = NULL, from_idx = NULL, to_idx = NULL)
}
\arguments{
\item{m}{a thmodel, with predictions}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
Plot pack temp, predicted pack temp, ambient temp, kW (smoothed)
}
\examples{
plot_fit(eNV200noac50kWh, from_idx = 1, to_idx = 10)
plot_fit(eNV200ac50kWh, "2026-01-26 11:00", "2026-01-27")
}
