% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gid_ahr.R
\name{plot_gid_ahr}
\alias{plot_gid_ahr}
\title{plot_gid_ahr: explore ahr v gid}
\usage{
plot_gid_ahr(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
Plots reveal that gid is a biased estimate of Ah remaining, with
a multiplicative factor of pack_volts explaining much of the variance.
}
\examples{
plot_gid_ahr(eNV200ac24kWh_2025)
plot_gid_ahr(eNV200ac24kWh_2025, min_soc = 70)
}
