% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gid_kWh.R
\name{plot_gid_kWh}
\alias{plot_gid_kWh}
\title{plot_gid_kWh: gids*.080, estimated kWh, temps, inefficiencies}
\usage{
plot_gid_kWh(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
plot_gid_kWh: gids*.080, estimated kWh, temps, inefficiencies
}
\examples{
plot_gid_kWh(eNV200ac24kWh_2025)
}
