% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kwh.R
\name{plot_kWh}
\alias{plot_kWh}
\title{plot_kwh: study gids -- do they report kWh or Ah?}
\usage{
plot_kWh(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  min_soc = NULL,
  max_soc = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{min_soc}{high-pass filter on soc (to examine non-linearity)}

\item{max_soc}{low-pass filter on soc (to examine non-linearity)}
}
\value{
an Environment
}
\description{
plot_kwh: study gids -- do they report kWh or Ah?
}
\examples{
plot_kWh(eNV200ac24kWh_2025)
plot_kWh(eNV200ac24kWh_2025, min_soc = 70)
}
