% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log.R
\name{plot_log}
\alias{plot_log}
\title{plot_log: plots distance, elevation, speed, SOC; geolocates in title}
\usage{
plot_log(m, from_date = NULL, to_date = NULL, from_idx = NULL, to_idx = NULL)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
plot_log: plots distance, elevation, speed, SOC; geolocates in title
}
\examples{
plot_log(predict_temp())
plot_log(predict_temp(), 1, 10)
}
