% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_ocv.R
\name{plot_log_ocv}
\alias{plot_log_ocv}
\title{plot_log_ocv: plots pack_volts-est_ocv, pack_amps, soc}
\usage{
plot_log_ocv(m, from_date = NULL, to_date = NULL)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}
}
\value{
an Environment
}
\description{
plot_log_ocv: plots pack_volts-est_ocv, pack_amps, soc
}
\examples{
plot_log_ocv(predict_temp())
}
