% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soc_gid.R
\name{plot_soc_gid}
\alias{plot_soc_gid}
\title{plot_soc_gid: plot an estimated SOC or the LeafSpy-reported SOC against gid}
\usage{
plot_soc_gid(
  m,
  suppress_outliers = FALSE,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  min_soc = NULL,
  max_soc = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{suppress_outliers}{if TRUE, don't plot the reported outliers}

\item{from_date}{starting date/time for plotting & analysis}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{min_soc}{high-pass filter on soc (to examine non-linearity)}

\item{max_soc}{low-pass filter on soc (to examine non-linearity)}
}
\value{
an Environment
}
\description{
plot_soc_gid: plot an estimated SOC or the LeafSpy-reported SOC against gid
}
\examples{
plot_soc_gid(eNV200ac24kWh_2025)
plot_soc_gid(eNV200ac24kWh_2025, max_soc_filter = 25)
}
