% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soc_gid_old.R
\name{plot_soc_gid_old}
\alias{plot_soc_gid_old}
\title{plot_soc_gid_old: exploratory study of soc v gid/soh}
\usage{
plot_soc_gid_old(
  m,
  suppress_outliers = TRUE,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  min_soc = NULL,
  max_soc = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{suppress_outliers}{if TRUE, won't plot the reported outliers}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{min_soc}{high-pass filter on soc (to examine non-linearity)}

\item{max_soc}{low-pass filter on soc (to examine non-linearity)}
}
\value{
an Environment
}
\description{
Possibly still useful for visualisation.
}
\examples{
plot_soc_gid_old(eNV200ac24kWh_2025)
plot_soc_gid_old(eNV200ac24kWh_2025, min_soc = 70)
}
