% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soh.R
\name{plot_soh}
\alias{plot_soh}
\title{plot_soh: plots SOH and Hx; geolocates in title}
\usage{
plot_soh(m, from_date = NULL, to_date = NULL, from_idx = NULL, to_idx = NULL)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
plot_soh: plots SOH and Hx; geolocates in title
}
\examples{
plot_soh(predict_temp())
plot_soh(predict_temp(), from_idx=1, to_idx=10)
}
