% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soh_effR.R
\name{plot_soh_effR}
\alias{plot_soh_effR}
\title{plot_soh_effR: plots SOH, Hx, pack temps, and estimated effR}
\usage{
plot_soh_effR(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
an Environment
}
\description{
plot_soh_effR: plots SOH, Hx, pack temps, and estimated effR
}
\examples{
plot_soh_effR(predict_temp())
plot_soh_effR(predict_temp(), from_idx=1, to_idx=10)
}
