% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temps.R
\name{plot_temps}
\alias{plot_temps}
\title{plot_temps: plots temperatures of all sensors, mean temp, and predicted temp}
\usage{
plot_temps(m)
}
\arguments{
\item{m}{a thmodel}
}
\value{
an Environment
}
\description{
plot_temps: plots temperatures of all sensors, mean temp, and predicted temp
}
\examples{
plot_temps(predict_temp())
}
