% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volts_gids.R
\name{plot_volts_gids}
\alias{plot_volts_gids}
\title{plot_volts_gids: scatterplot of pack_volts v scaled gids}
\usage{
plot_volts_gids(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  min_sgids = NULL,
  max_sgids = NULL,
  volts_per_cell = T,
  scale_colours = 0
)
}
\arguments{
\item{m}{a thmodel with predictions (requires m$pack_avg_temp)}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{min_sgids}{high-pass filter on scaled gids (to examine non-linearity)}

\item{max_sgids}{low-pass filter on scaled gids (to examine non-linearity)}

\item{scale_colours}{colours show pack_temp (0), pack_amps (1), pack_kW (2)}

\item{volts_per_cell:}{T to divide pack_volts by the number of cell(pair)s}
}
\value{
an Environment
}
\description{
plot_volts_gids: scatterplot of pack_volts v scaled gids
}
\examples{
plot_volts_gids(eNV200ac24kWh_2025)
plot_volts_gids(eNV200ac24kWh_2025, max_sgids = 75)
plot_volts_gids(eNV200ac24kWh_2025, max_sgids = 75, temp_colours=FALSE)
plot_volts_gids(eNV200ac24kWh_2025, max_sgids = 75, temp_colours=TRUE)
}
