% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volts_pred.R, R/plot_volts_pred_volts.R
\name{plot_volts_pred_volts}
\alias{plot_volts_pred_volts}
\title{plot_volts_pred_volts: plot of prediction error in pack_volts by pack_volts}
\usage{
plot_volts_pred_volts(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  from_temp = NULL,
  to_temp = NULL,
  max_amps = NULL,
  wonky_threshold = 50,
  by_amps = TRUE
)

plot_volts_pred_volts(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  from_temp = NULL,
  to_temp = NULL,
  max_amps = NULL,
  wonky_threshold = 50,
  by_amps = TRUE
)
}
\arguments{
\item{m}{a thmodel or an ocv_model}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in ocv_model, ignored if !is.null(from_date)}

\item{to_idx}{ending index in ocv_model, ignored if !is.null(to_date)}

\item{from_temp}{lower limit of battery temps to be analysed}

\item{to_temp}{upper limit of battery temps to be analysed}

\item{max_amps}{limiting amperage for plotted points}

\item{wonky_threshold}{in Volts, outlier criterion (default 50)}

\item{by_amps}{colours indicate pack_temp (F), pack_amps (T)}
}
\value{
an Environment

an Environment
}
\description{
plot_volts_pred_volts: plot of prediction error in pack_volts by pack_volts

plot_volts_pred_volts: plot of prediction error in pack_volts by pack_volts
}
\examples{
plot_volts_pred_volts(predict_volts_tafel(eNV50kWh))
plot_volts_pred_volts(predict_volts_tafel(eNV50kWh))
}
