% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_volts_tafel.R
\name{predict_volts_tafel}
\alias{predict_volts_tafel}
\title{Predicts pack_volts from an SOC lookup in ocv_tbl, conditioned by pack_amps,
hx, and pack_temperature.  Parameters include a polarisation "resistance" (in
a Tafel model).}
\usage{
predict_volts_tafel(
  tm = NULL,
  polarisation_irr = 60,
  tafel_slope = 0.5,
  lambda_polarisation = 120,
  arrhenius_tafel = -3500,
  soc_offset_corr = 0.35,
  soc_slope_corr = 0.8,
  ocv_tbl = NULL,
  trace = 2
)
}
\arguments{
\item{tm}{a thmodel}

\item{polarisation_irr}{in mOhms, baseline polarisation response}

\item{tafel_slope}{large-C polarisation response}

\item{lambda_polarisation}{in s, decay rate of polarisation}

\item{arrhenius_tafel}{in K, Arrhenius coefficient for tafel_slope}

\item{soc_offset_corr}{offset of actual SOC from LeafSpy-reported SOC}

\item{soc_slope_corr}{slope-correction for LeafSpy-reported SOC}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose, 3 with a plot}
}
\value{
a thmodel with volt-prediction columns in its logdata
}
\description{
Predicts pack_volts from an SOC lookup in ocv_tbl, conditioned by pack_amps,
hx, and pack_temperature.  Parameters include a polarisation "resistance" (in
a Tafel model).
}
\examples{
m <- predict_volts_tafel(eNV50kWh)
}
