% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmodel.R
\name{set_name_thmodel}
\alias{set_name_thmodel}
\title{setter for the name field of thmodel}
\usage{
set_name_thmodel(m, nm)
}
\arguments{
\item{m}{a thmodel}

\item{nm}{a friendly and informative name for this thmodel}
}
\value{
the modified thmodel
}
\description{
setter for the name field of thmodel
}
\examples{
m <- set_name_thmodel(predict_temp(), "26Jan2026, 50kWh, no AC")
}
