% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmodel.R
\name{default_params}
\alias{default_params}
\title{default_params: sets thmodel params to default values, if the parameter
list has length 0}
\usage{
default_params(m)
}
\arguments{
\item{m}{a thmodel}
}
\value{
a thmodel with all required params set
}
\description{
default_params: sets thmodel params to default values, if the parameter
list has length 0
}
\examples{
default_params(new_thmodel())
}
