% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eNV200ac50kWh}
\alias{eNV200ac50kWh}
\title{eNV200ac50kWh: a thmodel object, with provenance, of logs from the
fully-functional 50kWh upgrade.  The fields are as defined in thmodel.R.}
\format{
A thmodel object.
}
\source{
{thmodel} R package.
}
\usage{
eNV200ac50kWh
}
\description{
eNV200ac50kWh: a thmodel object, with provenance, of logs from the
fully-functional 50kWh upgrade.  The fields are as defined in thmodel.R.
}
\keyword{datasets}
