% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_ocv.R
\name{est_ocv}
\alias{est_ocv}
\title{est_ocv: refine an ocv_tbl, from errors in an om_model's voltage prediction}
\usage{
est_ocv(om, wonky_threshold = 50, trace = 2)
}
\arguments{
\item{om}{an ocv_model}

\item{wonky_threshold}{in Volts, outlier criterion (default 50)}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}
}
\value{
an ocv_model with an updated ocv_tbl and voltage predictions
}
\description{
est_ocv: refine an ocv_tbl, from errors in an om_model's voltage prediction
}
\examples{
est_ocv(om_eNV50kWh)
}
