% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model: uses nlm() to find a best-fit, with MSE criterion}
\usage{
fit_model(
  m = NULL,
  effective_pack_resistance = NA,
  lambda_cell_to_pack = NA,
  lambda_pack_to_ambient = NA,
  lambda_pack_AC_to_ambient = NA,
  COP = NA,
  print.level = 1
)
}
\arguments{
\item{m}{a thmodel}

\item{effective_pack_resistance}{in mOhms}

\item{lambda_cell_to_pack}{in seconds}

\item{lambda_pack_to_ambient}{in hours}

\item{lambda_pack_AC_to_ambient}{in hours}

\item{COP}{coefficient of heatpump performance, dimensionless}

\item{print.level}{0 for quiet, 1 for start and endpoints, 2 for verbose}

\item{x}{parameter list}
}
\value{
a list, retval from nlm() describing its best-fit

MSE of the fit
}
\description{
fit_model: uses nlm() to find a best-fit, with MSE criterion
}
\examples{
m <- fit_model(print.level = 2)
fm: interface to predict_temp(), for use by nlm()

Side effect: updates thmodel `m` in the calling environment

fm(c(0.4,120,8,2,3))
}
