% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model: uses nlm() to find a best-fit, with MSE criterion}
\usage{
fit_model(m = NULL, packr = 4, lambda1 = 100, lambda2 = 300, print.level = 1)
}
\arguments{
\item{m}{a thmodel}

\item{packr}{starting value of effective resistance, in Ohms}

\item{lambda1}{starting value of the cell-to-module time constant, in seconds}

\item{lambda2}{starting value of the module-to-ambient time constant, in sec}

\item{print.level}{0 for quiet, 1 for start and endpoints, 2 for verbose}

\item{x}{parameter list}
}
\value{
list retval from nlm(), describing its best-fit

MSE of the fit
}
\description{
fit_model: uses nlm() to find a best-fit, with MSE criterion
}
\examples{
m <- fit_model(print.level = 2)
fm: interface to predict_temp(), for use by nlm()

Side effect: updates thmodel `m` in the calling environment

fm(c(1,100,1000))
}
