% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_r_to_ocv.R
\name{fit_r_to_ocv}
\alias{fit_r_to_ocv}
\title{fit_r_to_ocv: uses optim() to find a best-fit pack resistance to an ocv_model}
\usage{
fit_r_to_ocv(om = new_ocv_model(), iter_count = 4, trace = 1)
}
\arguments{
\item{om}{an ocv_model}

\item{iter_count}{controls convergence on predicted temps}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}

\item{x}{parameter list}
}
\value{
modified ocv_model, with best-fit resistive params
}
\description{
fit_r_to_ocv: uses optim() to find a best-fit pack resistance to an ocv_model
}
\examples{
om <- fit_r_to_ocv(new_ocv_model("eNV50kWh", list(eNV200ac50kWh, eNV200noac50kWh)))
fom: local fcn, for use by optim()

Side effect: updates ocv_model om in the calling environment

}
