% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{MSE_of_fit}
\alias{MSE_of_fit}
\title{MSE_of_fit: compute the mean squared error of the estimation in a logtibble}
\usage{
MSE_of_fit(m)
}
\arguments{
\item{m, }{a thmodel}
}
\value{
double
}
\description{
MSE_of_fit: compute the mean squared error of the estimation in a logtibble
}
\examples{
MSE_of_fit(predict_temp())
}
