% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_of_ocv_fit.R
\name{MSE_of_ocv_fit}
\alias{MSE_of_ocv_fit}
\title{MSE_of_ocv_fit: compute the mean squared error of the estimation}
\usage{
MSE_of_ocv_fit(om)
}
\arguments{
\item{om, }{an ocv_model}
}
\value{
double
}
\description{
MSE_of_ocv_fit: compute the mean squared error of the estimation
}
\examples{
MSE_of_ocv_fit(
   predict_ocv(new_ocv_model("eNV50kWh", list(eNV200ac50kWh)), 150, -3500)))
}
