% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_logfile.R
\name{munge_logfile}
\alias{munge_logfile}
\title{Munges a csv to ensure it has no VIN column, has timestamps in ISO format,
and has tidied column names.  Side effect: updates this csv file.  Returns
a thmodel with this csv as its logdata, and with provenance fields set.}
\usage{
munge_logfile(
  logfildir = "data-raw",
  logfilnm = "log26Jan2026.csv",
  logname = "26Jan2026 50kWh noAC"
)
}
\arguments{
\item{logfildir}{directory of csv logfile to be interpreted as a thmodel}

\item{logfilnm}{name of logfile to be interpreted}

\item{logname}{friendly and informative name for this logfile}
}
\value{
thmodel
}
\description{
Munges a csv to ensure it has no VIN column, has timestamps in ISO format,
and has tidied column names.  Side effect: updates this csv file.  Returns
a thmodel with this csv as its logdata, and with provenance fields set.
}
\examples{
m <- munge_logfile()
}
