% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_logfile.R
\name{munge_logfile}
\alias{munge_logfile}
\title{Munges a csv to ensure it has no VIN column, has timestamps in ISO format,
is sorted on these timestamps, and has tidied column names.  Side effect:
updates this csv file.  Returns a thmodel with the munged csv as its logdata,
and with provenance fields set.}
\usage{
munge_logfile(
  logfilnm = "log26Jan2026.csv",
  logfildir = "data-raw",
  logname = NULL,
  USonian_dates = FALSE
)
}
\arguments{
\item{logfilnm}{name of logfile to be interpreted}

\item{logfildir}{directory of csv logfile, optional, default is "data-raw"}

\item{logname}{friendly and informative name for this logfile, optional}

\item{USonian_dates}{TRUE: mdy parsing; FALSE (default): dmy parsing}
}
\value{
thmodel
}
\description{
Todo: munge the filenm, if it contains information from a VIN.  Perhaps:
allow the user to specify a pseudoVIN, to retain provenance informmation
on csv files collected from multiple vehicles.
}
\examples{
m <- munge_logfile()
}
