% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ocv_model.R
\name{new_ocv_model}
\alias{new_ocv_model}
\title{new_ocv_model: S3 class constructor}
\usage{
new_ocv_model(name = "", thmodels = list())
}
\arguments{
\item{name}{# user-friendly name, used for titles of plots}

\item{thmodels}{# provenance (a list of thmodels)}
}
\value{
ocv_model, a named list
}
\description{
new_ocv_model: S3 class constructor
}
\examples{
om <- new_ocv_model("eNV50kWh", list(eNV200ac50kWh, eNV200noac50kWh))
}
