% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ocv_model.R
\name{new_ocv_model}
\alias{new_ocv_model}
\title{A lean-and-mean constructor, defining only the names and classes of the
required fields.}
\usage{
new_ocv_model()
}
\value{
ocv_model, a named list
}
\description{
A lean-and-mean constructor, defining only the names and classes of the
required fields.
}
\examples{
o <- new_ocv_model()
}
