% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmodel.R
\name{new_thmodel}
\alias{new_thmodel}
\title{new_thmodel.  A lean-and-mean constructor, defining only the names and
classes of the required fields.}
\usage{
new_thmodel()
}
\value{
thmodel, a named list
}
\description{
new_thmodel.  A lean-and-mean constructor, defining only the names and
classes of the required fields.
}
\examples{
m <- new_thmodel()
}
