% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot pack temp, predicted pack temp, ambient temp, pack kW}
\usage{
plot_fit(m)
}
\arguments{
\item{m}{}
}
\description{
Not working yet -- pack kW should be on a separate panel, or smoothed
so that it's not so intrusive
}
\details{
#todo: dispatch through generic plot()
}
\examples{
plot_fit(predict_temp())
plot_fit(fit_model())
}
