% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot pack temp, predicted pack temp, ambient temp, kW (smoothed)}
\usage{
plot_fit(m)
}
\arguments{
\item{m}{}
}
\value{
1
}
\description{
#todo: dispatch through generic plot()
}
\examples{
plot_fit(predict_temp())
plot_fit(fit_model())
}
