% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gids.R
\name{plot_gids}
\alias{plot_gids}
\title{plot_gids: scatterplot of SOC and pack_volts v gids, also a timeseries plot}
\usage{
plot_gids(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  max_soc = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{max_soc}{low-pass filter on soc (to examine non-linearity)}
}
\value{
an Environment
}
\description{
plot_gids: scatterplot of SOC and pack_volts v gids, also a timeseries plot
}
\examples{
plot_gids(eNV200ac24kWh_2025)
plot_gids(eNV200ac24kWh_2025, min_soc = 25)
}
