% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log.R
\name{plot_log}
\alias{plot_log}
\title{plot_log: plots distance, elevation, speed, SOC; geolocates in title}
\usage{
plot_log(m)
}
\arguments{
\item{m}{a thmodel}
}
\value{
an Environment
}
\description{
plot_log: plots distance, elevation, speed, SOC; geolocates in title
}
\examples{
plot_log(predict_temp())
}
