% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volts.R
\name{plot_volts}
\alias{plot_volts}
\title{plot_gids: scatterplot of pack_volts v scaled gids}
\usage{
plot_volts(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  max_soc = NULL
)
}
\arguments{
\item{m}{a thmodel}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{max_sgids}{low-pass filter on scaled gids (to examine non-linearity)}
}
\value{
an Environment
}
\description{
plot_gids: scatterplot of pack_volts v scaled gids
}
\examples{
plot_volts(eNV200ac24kWh_2025)
plot_volts(eNV200ac24kWh_2025, min_sgids = 25)
}
