% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volts_pred.R
\name{plot_volts_pred}
\alias{plot_volts_pred}
\title{plot_volts_pred: plot of soc by prediction error in pack_volts}
\usage{
plot_volts_pred(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  wonky_threshold = 50,
  scatter = FALSE,
  scale_colours = 0
)
}
\arguments{
\item{m}{a thmodel or an ocv_model}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in ocv_model, ignored if !is.null(from_date)}

\item{to_idx}{ending index in ocv_model, ignored if !is.null(to_date)}

\item{wonky_threshold}{in Volts, outlier criterion (default 50)}

\item{scatter}{TRUE for scatterplot, FALSE for box and whiskers}

\item{scale_colours}{colours indicate pack_temp (0), pack_amps (1)}
}
\value{
an Environment
}
\description{
plot_volts_pred: plot of soc by prediction error in pack_volts
}
\examples{
plot_volts_pred(om_eNV50kWh)
}
