% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volts_ts.R
\name{plot_volts_ts}
\alias{plot_volts_ts}
\title{plot_volts_ts: plot time-series of pack_volts, pred_pack_volts_from_soc}
\usage{
plot_volts_ts(
  m,
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL,
  min_sgids = NULL,
  max_sgids = NULL,
  volts_per_cell = T,
  scale_colours = 0
)
}
\arguments{
\item{m}{a thmodel with predictions from predict_volts_tafel()}

\item{from_date}{starting date/time}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}

\item{volts_per_cell:}{T to divide pack_volts by the number of cell(pair)s}
}
\value{
a ggplot() environment
}
\description{
plot_volts_ts: plot time-series of pack_volts, pred_pack_volts_from_soc
}
\examples{
plot_volts_ts(predict_volts_tafel(eNV50kWh))
}
