% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_temp.R
\name{predict_temp}
\alias{predict_temp}
\title{Evaluates a thermal model to predict temperatures in a LeafSpy log}
\usage{
predict_temp(
  tmodel = NULL,
  effective_pack_resistance = NA,
  packr85 = NA,
  polarisation_energy = NA,
  polarisation_heat = NA,
  lambda_module_to_ambient = NA,
  lambda_module_AC_to_ambient = NA,
  fan_power = NA,
  COP = NA,
  arrhenius_resistance = NA,
  heat_capacity = NA,
  ocv_tbl = NULL,
  iter_count = 4,
  min_segment_length = 20,
  trace = 2,
  logfilnm = "log26Jan2026.csv",
  logfildir = "data-raw",
  from_date = NULL,
  to_date = NULL,
  from_idx = NULL,
  to_idx = NULL
)
}
\arguments{
\item{tmodel}{a thmodel, optional}

\item{effective_pack_resistance}{in mOhms at 298.15K for SOC <= 70 percent}

\item{packr85}{in mOhms, effective pack resistance at SOC = 85 percent}

\item{polarisation_energy}{in kJ/V, a reversible (entropic) heat}

\item{polarisation_heat}{in J/(ΔI)², enthalpy of a polarisation shift}

\item{lambda_module_to_ambient}{in hours}

\item{lambda_module_AC_to_ambient}{in hours}

\item{fan_power}{in Watts}

\item{COP}{dimensionless}

\item{arrhenius_resistance}{in K, temperature dependence of packr}

\item{heat_capacity}{in kJ/K}

\item{ocv_tbl}{maps SOC onto OCV, either a 2-column tibble or an om_model}

\item{iter_count}{may be increased for a more accurate prediction}

\item{min_segment_length}{shorter sequences of samples are ignored}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}

\item{logfilnm}{name of a csv logfile to be read, if is.null(tmodel)}

\item{logfildir}{directory in which the logfile is located}

\item{from_date}{starting date/time for calculation of MSE}

\item{to_date}{ending date/time}

\item{from_idx}{starting index in thmodel, ignored if !is.null(from_date)}

\item{to_idx}{ending index in thmodel, ignored if !is.null(to_date)}
}
\value{
a thmodel
}
\description{
A LeafSpy logfile may be specified to this function by filename and
directory, in which case this logfile is munged -- to mitigate the privacy
risk of publishing a VIN, and to revise column names so that they're tidy.
This routine also does some "cleaning" of obviously-wonky data e.g. of a
pack at 0 Volts or at 80 degrees (in a field that is normally in Centrigade
units).
}
\details{
The output of this routine is a thmodel object containing the munged and
cleaned data from LeafSpy augmented with additional columns for the
predictions of pack temperature (and for convenience when plotting). This
thmodel records a timestamp of its modification, and the values of the
modelling parameters used for its updated temperature predictions.

Initial estimates of parameters are hardcoded in default_params().
}
\examples{
m <- predict_temp()
m <- predict_temp(m, effective_pack_resistance = 67)
}
