% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_temp.R
\name{predict_temp}
\alias{predict_temp}
\title{Predict pack temperatures using a simple thermal model and data from
a LeafSpy logfile.}
\usage{
predict_temp(
  effective_pack_resistance = 4,
  lambda_cell_to_pack = 60,
  lambda_pack_to_ambient = NULL,
  cell_heat_capacity = 1e+07,
  pack_heat_capacity = NULL,
  logfile = NULL,
  logtibble = NULL
)
}
\arguments{
\item{effective_pack_resistance}{in Ohms}

\item{lambda_cell_to_pack}{time constant, in seconds}

\item{lambda_pack_to_ambient}{time constant, in seconds}

\item{cell_heat_capacity}{in J/gK, total over all modules in pack}

\item{pack_heat_capacity}{in J/gK, of pack excluding modules and cells}

\item{logfile}{name of logfile (in .csv format)}

\item{logtibble}{required input, if logfile == "".  Superassignment
to a global tibble with this name, if logfile <> ""}
}
\value{
xts with columns pack_avg_temp, pred_pack_avg_temp, soc,
  pack_amps, charging_kW, pack_kW
}
\description{
Predict pack temperatures using a simple thermal model and data from
a LeafSpy logfile.
}
\examples{
predict_temp()
}
