% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_volts.R
\name{predict_volts}
\alias{predict_volts}
\title{Predicts pack_volts from OCV(SOC), pack_amps, hx, pack_temperature, arr}
\usage{
predict_volts(
  om = NULL,
  effective_pack_resistance = NA,
  arrhenius_resistance = NA,
  packr85 = NA,
  trace = 1
)
}
\arguments{
\item{om}{an ocv_model}

\item{effective_pack_resistance}{in mOhms at 298.15K}

\item{arrhenius_resistance}{in K}

\item{packr85}{in mOhms at 298.15K, at SOC = 0.85 (inflection at soc = 0.7)}

\item{trace}{0 for silent, 1 for minimal, 2 for verbose}
}
\value{
a ocv_model, with a pred_pack_volts column in its logdata
}
\description{
Predicts pack_volts from OCV(SOC), pack_amps, hx, pack_temperature, arr
}
\examples{
om <- predict_volts(om_eNV50kWh)
}
