% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thmodel_from_directory.R
\name{thmodel_from_directory}
\alias{thmodel_from_directory}
\title{Utility to import an entire directory of csv files into a thmodel}
\usage{
thmodel_from_directory(
  logfildir = "data-raw/eNV200noac50kWh",
  name = NULL,
  USonian_dates = FALSE,
  capacity = NULL
)
}
\arguments{
\item{logfildir}{path to directory containing csv files}

\item{name}{friendly name for this thmodel}

\item{USonian_dates}{TRUE for mdy, FALSE for dmy}

\item{capacity}{nominal capacity of the pack, in kWh}
}
\value{
a thmodel
}
\description{
Warning: any update to LeafSpy may cause it to produce logfiles with a
different set of column names, so some additional data-tidying may be
required.
}
\details{
warning: this script munges the csv files in the target directory, using
'munge_logfile()'
}
\examples{
m <- thmodel_from_directory()
}
